Adding profiles blueprint
